chore: 사용하지 않는 요소 제거
