+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Scene/CoinDetail/CoinDetailState.swift
@@ -5,12 +5,6 @@
 //  Created by Ryan-Son on 2022/03/12.
 //
 
-extension String {
-    func symbolToName() -> String {
-        return self.replacingOccurrences(of: "_KRW", with: "")
-    }
-}
-
 struct CoinDetailState: Equatable {
     let symbol: String
 }
