+++ CryptocurrencyExchange/CryptocurrencyExchange/Sources/Presentations/Scene/CoinDetail/Component/CurrentTicker/CoinCurrentTickerAction.swift
@@ -10,6 +10,6 @@ import Foundation
 enum CoinCurrentTickerAction: Equatable {
     case onAppear
     case onDisappear
-    case responseTickerSingle(Result<BithumbTickerSingle, OrderBookListError>)
-    case responseTickerStream(Result<BithumbTickerStream, OrderBookListError>)
+    case responseTickerSingle(Result<BithumbTickerSingle, CoinCurrentTickerError>)
+    case responseTickerStream(Result<BithumbTickerStream, CoinCurrentTickerError>)
 }
